move cluster reducer out to its own file
